вынес функции в utils переименовал колбэк события по критерию Д9
