+++ src/mocks/offers.ts
@@ -2763,16 +2763,3 @@ export const mockOffers: Offers = [
   }
 ];
 
-export const getFavoriteOffers = (): Offers => mockOffers.filter((offer) => offer.isFavorite).sort((offer1, offer2) => (offer1.city.name > offer2.city.name) ? 1 : -1);
-
-export const getOffersByCity = (offers: Offers) => {
-  const offersByCity: {city: string; offers: Offers}[] = [];
-  let cityName: string = '';
-  offers.map((offer) => {
-    if (offer.city.name !== cityName) {
-      offersByCity.push({city: offer.city.name, offers: offers.filter((item) => item.city.name === offer.city.name)});
-    }
-    cityName = offer.city.name;
-  });
-  return offersByCity;
-};
